개발/트렌드분석대시보드/0329_trend_html최신본/requirements.txt
Flask==2.2.3
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.3
MarkupSafe==2.1.2
python-dotenv==1.0.0
mysql-connector-python==8.0.33
pandas==1.5.3
numpy==1.24.3
matplotlib==3.7.1
networkx==3.1
wordcloud==1.9.1
scikit-learn==1.2.2
flask-caching==2.0.2
pillow==9.5.0